Initial Upload
This file is a list of members of a set that are listed in arbitrary order, instead of lexicographical order.  There are a total of 1.89321481108517289984e+20 member of this set.  The question is, can any person devise software that is able to predict the order in which future members of the set will be produced, given some even large number of the member listed in the corresponding text file?